+++ terraplanfeed/__init__.py
@@ -1,2 +1,10 @@
-"""Main package for terraplanfeed."""
-__version__ = "0.1.2"
+"""
+Main package for terraplanfeed.
+
+This package receives a file in Terraform JSON output, parses to get the
+changes, and creates output that can be sent to different output drivers.
+
+Modules:
+    terraform
+"""
+__version__ = "0.2.0"