+++ CHANGELOG.md
@@ -3,6 +3,32 @@
 ## [Unreleased]
 
 
+<a name="0.2.0"></a>
+## [0.2.0] - 2021-01-30
+### Chore
+- vscode settings for project
+
+### Docs
+- docstrings update
+- terraform link for json docs
+
+### Feat
+- create resource attributes
+- terraform functions for filtering data
+- module structure
+- parse terraform plan
+- **terraform:** calculate resource name
+
+### Fix
+- validate json file
+
+
+<a name="0.1.3"></a>
+## [0.1.3] - 2021-01-25
+### Fix
+- gitignore
+
+
 <a name="0.1.2"></a>
 ## [0.1.2] - 2021-01-25
 ### Chore
@@ -11,6 +37,9 @@
 ### Fix
 - requirements.txt
 - setup.py classifiers
+
+### Pull Requests
+- Merge pull request [#3](https://github.com/bcochofel/terraplanfeed/issues/3) from bcochofel/env_configuration
 
 
 <a name="0.1.1"></a>
@@ -43,6 +72,8 @@
 - Merge pull request [#1](https://github.com/bcochofel/terraplanfeed/issues/1) from bcochofel/package_skel
 
 
-[Unreleased]: https://github.com/bcochofel/terraplanfeed/compare/0.1.2...HEAD
+[Unreleased]: https://github.com/bcochofel/terraplanfeed/compare/0.2.0...HEAD
+[0.2.0]: https://github.com/bcochofel/terraplanfeed/compare/0.1.3...0.2.0
+[0.1.3]: https://github.com/bcochofel/terraplanfeed/compare/0.1.2...0.1.3
 [0.1.2]: https://github.com/bcochofel/terraplanfeed/compare/0.1.1...0.1.2
 [0.1.1]: https://github.com/bcochofel/terraplanfeed/compare/0.1.0...0.1.1